setting up with docker image
