+++ .dockerignore
@@ -1,0 +1,6 @@
+node_modules
+client/node_modules
+server/node_modules
+package-lock.json
+client/package-lock.json
+server/package-lock.json
