+++ client/index.html
@@ -9,6 +9,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/main.jsx"></script>
+    <script type="module" src="/src/Main.jsx"></script>
   </body>
 </html>
