+++ client/src/authService.js
@@ -1,0 +1,39 @@
+import {
+  createUserWithEmailAndPassword,
+  signInWithEmailAndPassword,
+  signOut,
+} from "firebase/auth"
+import { auth } from "../../../server/firebase"
+
+// Sign up new user
+export async function signUp(email, password) {
+  try {
+    const userCredential = await createUserWithEmailAndPassword(
+      auth,
+      email,
+      password
+    )
+    return userCredential.user
+  } catch (error) {
+    throw error
+  }
+}
+
+// Sign in existing user
+export async function signIn(email, password) {
+  try {
+    const userCredential = await signInWithEmailAndPassword(
+      auth,
+      email,
+      password
+    )
+    return userCredential.user
+  } catch (error) {
+    throw error
+  }
+}
+
+// Sign out
+export async function logOut() {
+  await signOut(auth)
+}
