+++ server/firebaseConfig.json
@@ -1,0 +1,9 @@
+{
+  "apiKey": "AIzaSyAGO2bh8PECz-W9lR37jx0R3oEt2xsawF8",
+  "authDomain": "astuteabroad-0525.firebaseapp.com",
+  "projectId": "astuteabroad-0525",
+  "storageBucket": "astuteabroad-0525.firebasestorage.app",
+  "messagingSenderId": "167193485736",
+  "appId": "1:167193485736:web:ce2d766a886665b6317d9d",
+  "measurementId": "G-0WEP1RY3XB"
+}
